Add files via upload
Objective
The aim of this task was to independently explore the provided data files, which include students' performance on STAAR EOC for Algebra I and student demographic information. The goal was to merge the data sets, perform an exploratory analysis, and communicate the findings via PowerPoint slides. This assignment was designed to evaluate proficiency in data cleaning, analysis, and visualization techniques.